Added authorization via simple password.
The password is provided to the script on the command line via the parameter --key. The string "none" (default) means that no password is required.
Clients pass the password as an URL parameter, e.g. http://localhost:8049/?key=<mypassword>.
The server will respond with code 401 (unauthorized) when the provided key is incorrect.